First registration shouldn't block the application using SBA
When pyctuator is starting up, it starts sending periodic registration
requests to SBA. Currently, the first registration is done on the "main
thread" while subsequent registrations are done asynchronously (using
a "daemonic" `threading.Timer()`.

If the registration's HTTP request is stuck until timing out, this blocks
the application until the this timeout.

Instead, all registration attempts, including the first one, should be
done asynchronously.

See #51

+++ pyctuator/pyctuator.py
@@ -33,7 +33,7 @@ class Pyctuator:
             registration_url: Optional[str],
             registration_auth: Optional[Auth] = None,
             app_description: Optional[str] = None,
-            registration_interval_sec: int = 10,
+            registration_interval_sec: float = 10,
             free_disk_space_down_threshold_bytes: int = 1024 * 1024 * 100,
             logfile_max_size: int = 10000,
             logfile_formatter: str = default_logfile_format,
